+++ R/lab_import_checkr.R
@@ -22,13 +22,12 @@ lab_import_checkr <- function(x) {
 #' Default method for lab_import_checkr()
 #'
 #' Allow data frames with unspecified object class to pass through
-#' @param x data frame containing variables
-#' @param ... further arguments passed to or from other methods
+#' @inheritParams lab_import_checkr
 #' @author Jay Achar \email{jay.achar@@doctors.org.uk}
 #' @seealso \code{\link{tbcleanr}}
 #' @export
 
-lab_import_checkr.default <- function(x, ...) {
+lab_import_checkr.default <- function(x) {
 
   message("No lab object class detected: lab_import_checkr() not applied.")
   x