+++ R/lab_subset.R
@@ -25,8 +25,7 @@ lab_subset <- function(x, add = NULL, ...) {
 #' Default method for lab_subset()
 #'
 #' Allow data frames with unspecified object class to pass through
-#' @param x data frame containing variables
-#' @param ... further arguments passed to or from other methods
+#' @inheritParams lab_subset
 #' @author Jay Achar \email{jay.achar@@doctors.org.uk}
 #' @seealso \code{\link{tbcleanr}}
 #' @export