+++ R/lab_subset.koch6.R
@@ -1,10 +1,8 @@
 #' Subset lab variables
 #'
 #' Subset pre-specified TB laboratory variables
-#' @param x data frame containing variables
-#' @param add string of any additional variables to keep
-#' @param ... further arguments passed to or from other methods
-#' @author Jay Achar \email{jay.achar@@doctors.org.uk}
+#' @inheritParams lab_subset
+#' @author Jay Achar 
 #' @seealso \code{\link{tbcleanr}}
 #' @export 
 
@@ -14,7 +12,7 @@ lab_subset.koch6 <- function(x, add = NULL, ...) {
   object_class <- class(x)
   
   # define variables to keep
-  keep <- c("RegistrationNb",
+  keep <- c("RegistrationNb", "Id_Bacterio",
             "Samplecollectiondate", 
             "SmearResult", "CultResult", 
             "GeneXpertResult", "HainResult", "DSTLabNumber1",