+++ tests/testthat/test-lab_date_fixer.R
@@ -34,3 +34,18 @@ test_that("Grozny works fine", {
 test_that("Errors, warnings and messages", {
   expect_message(data.frame(x = 2) %>% lab_date_fixer(), "No lab object class")
 })
+
+
+## Koch 6
+k6_raw <- system.file("testdata", "lab_date_fixer_k6.rds", package = "tbcleanr") %>% 
+  readRDS()
+
+k6_data <- lab_date_fixer(k6_raw)
+
+test_that("Koch6 works fine", {
+  expect_equal(ncol(k6_data), 1L)
+  expect_equal(names(k6_data), "samp_date")
+  expect_equal(class(k6_data), class(k6_raw))
+  expect_equal(class(k6_data$samp_date), "Date")
+  expect_equal(k6_data$samp_date[1], lubridate::ymd("2012-01-01"))
+})